// apps/coinet-platform/src/l13/contracts/l13-constitutional-types.ts
export enum L13AllowedCapability {
  EXPLAIN_ENGINE_STATE = 'EXPLAIN_ENGINE_STATE',
  SUMMARIZE_MARKET_STATE = 'SUMMARIZE_MARKET_STATE',
  ANSWER_USER_QUESTION = 'ANSWER_USER_QUESTION',
  WRITE_ALERT_TEXT = 'WRITE_ALERT_TEXT',
  WRITE_STRUCTURED_REPORT = 'WRITE_STRUCTURED_REPORT',
  COMPARE_ASSETS = 'COMPARE_ASSETS',
  COMPARE_THESES = 'COMPARE_THESES',
  EXPLAIN_SCENARIOS = 'EXPLAIN_SCENARIOS',
  EXPLAIN_SCORES = 'EXPLAIN_SCORES',
  EXPLAIN_HYPOTHESES = 'EXPLAIN_HYPOTHESES',
  EXPLAIN_REGIME = 'EXPLAIN_REGIME',
  EXPLAIN_SEQUENCE = 'EXPLAIN_SEQUENCE',
  DISCLOSE_CONTRADICTION = 'DISCLOSE_CONTRADICTION',
  DISCLOSE_UNCERTAINTY = 'DISCLOSE_UNCERTAINTY',
  CITE_EVIDENCE_REFS = 'CITE_EVIDENCE_REFS',
  RESPECT_RESTRICTIONS = 'RESPECT_RESTRICTIONS',
  ADAPT_TONE_AND_LANGUAGE = 'ADAPT_TONE_AND_LANGUAGE',
  REFUSE_UNSUPPORTED_CONCLUSION = 'REFUSE_UNSUPPORTED_CONCLUSION',
}
// ...
export enum L13ForbiddenAction {
  INVENT_MISSING_SUPPORT = 'INVENT_MISSING_SUPPORT',
  HIDE_CONTRADICTION = 'HIDE_CONTRADICTION',
  IGNORE_RESTRICTION = 'IGNORE_RESTRICTION',
  OVERRIDE_CONFIDENCE_CAP = 'OVERRIDE_CONFIDENCE_CAP',
  REBUILD_SCENARIO = 'REBUILD_SCENARIO',
  REBUILD_SCORE = 'REBUILD_SCORE',
  REBUILD_HYPOTHESIS = 'REBUILD_HYPOTHESIS',
  REBUILD_SEQUENCE = 'REBUILD_SEQUENCE',
  REBUILD_REGIME = 'REBUILD_REGIME',
  CREATE_NEW_SCENARIO = 'CREATE_NEW_SCENARIO',
  CREATE_NEW_HYPOTHESIS = 'CREATE_NEW_HYPOTHESIS',
  COMPUTE_SCORE_LOCALLY = 'COMPUTE_SCORE_LOCALLY',
  TREAT_SCENARIO_CONFIDENCE_AS_PROBABILITY =
    'TREAT_SCENARIO_CONFIDENCE_AS_PROBABILITY',
  EMIT_BUY_INSTRUCTION = 'EMIT_BUY_INSTRUCTION',
  EMIT_SELL_INSTRUCTION = 'EMIT_SELL_INSTRUCTION',
  EMIT_HOLD_INSTRUCTION = 'EMIT_HOLD_INSTRUCTION',
  EMIT_AVOID_INSTRUCTION = 'EMIT_AVOID_INSTRUCTION',
  EMIT_LEVERAGE_INSTRUCTION = 'EMIT_LEVERAGE_INSTRUCTION',
  EMIT_POSITION_SIZE_INSTRUCTION = 'EMIT_POSITION_SIZE_INSTRUCTION',
  EMIT_ENTRY_EXIT_INSTRUCTION = 'EMIT_ENTRY_EXIT_INSTRUCTION',
  CLAIM_GUARANTEE = 'CLAIM_GUARANTEE',
  CLAIM_CERTAINTY_UNSUPPORTED = 'CLAIM_CERTAINTY_UNSUPPORTED',
  CALL_SCENARIO_WINNER = 'CALL_SCENARIO_WINNER',
  CALL_HYPOTHESIS_FINAL_TRUTH = 'CALL_HYPOTHESIS_FINAL_TRUTH',
  SAY_SCORE_MEANS_RECOMMENDATION = 'SAY_SCORE_MEANS_RECOMMENDATION',
  PRETEND_MISSING_DATA_COMPLETE = 'PRETEND_MISSING_DATA_COMPLETE',
  OMIT_ACTIVE_INVALIDATION = 'OMIT_ACTIVE_INVALIDATION',
  OMIT_REQUIRED_TRIGGER = 'OMIT_REQUIRED_TRIGGER',
  USE_RAW_LOWER_LAYER_BYPASS = 'USE_RAW_LOWER_LAYER_BYPASS',
  OUTPUT_UNGROUNDED_CLAIM = 'OUTPUT_UNGROUNDED_CLAIM',
}
// ...
export enum L13OutputSurfaceClass {
  AI_EXPLANATION_OUTPUT = 'AI_EXPLANATION_OUTPUT',
  AI_QUESTION_ANSWER_OUTPUT = 'AI_QUESTION_ANSWER_OUTPUT',
  AI_ALERT_OUTPUT = 'AI_ALERT_OUTPUT',
  AI_STRUCTURED_REPORT_OUTPUT = 'AI_STRUCTURED_REPORT_OUTPUT',
  AI_COMPARISON_OUTPUT = 'AI_COMPARISON_OUTPUT',
  AI_UNCERTAINTY_DISCLOSURE_OUTPUT = 'AI_UNCERTAINTY_DISCLOSURE_OUTPUT',
  AI_CONTRADICTION_DISCLOSURE_OUTPUT =
    'AI_CONTRADICTION_DISCLOSURE_OUTPUT',
  AI_BLOCKED_OUTPUT = 'AI_BLOCKED_OUTPUT',
  AI_GROUNDING_AUDIT_OUTPUT = 'AI_GROUNDING_AUDIT_OUTPUT',
}
// ...
export enum L13DownstreamConsumer {
  USER_FACING_DELIVERY = 'USER_FACING_DELIVERY',
  L14_DELIVERY = 'L14_DELIVERY',
  L15_FEEDBACK = 'L15_FEEDBACK',
  L16_GOVERNANCE = 'L16_GOVERNANCE',
  INTERNAL_AUDIT = 'INTERNAL_AUDIT',
}